modules/location-grpc/requirements.txt
grpcio==1.27.2
grpcio-tools==1.27.2
GeoAlchemy2==0.8.4
kafka-python
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
SQLAlchemy==1.3.19
shapely==1.7.0
python-dateutil==2.8.1
psycopg2-binary==2.8.5
